MAINT: consolidate _helpers and _util
