--- torch_np/_helpers.py
@@ -1,26 +1,0 @@
-import torch
-
-from . import _dtypes_impl, _util
-
-
-def ndarrays_to_tensors(*inputs):
-    """Convert all ndarrays from `inputs` to tensors. (other things are intact)"""
-    from ._ndarray import asarray, ndarray
-
-    if len(inputs) == 0:
-        return ValueError()
-    elif len(inputs) == 1:
-        input_ = inputs[0]
-        if isinstance(input_, ndarray):
-            return input_.tensor
-        elif isinstance(input_, tuple):
-            result = []
-            for sub_input in input_:
-                sub_result = ndarrays_to_tensors(sub_input)
-                result.append(sub_result)
-            return tuple(result)
-        else:
-            return input_
-    else:
-        assert isinstance(inputs, tuple)  # sanity check
-        return ndarrays_to_tensors(inputs)
